server/requirements.txt
ollama==0.1.6
python-dotenv
langchain-community==0.0.19